requirements.txt
torch>=2.0
transformers>=4.35
datasets>=2.14
scikit-learn>=1.3
matplotlib>=3.7
numpy>=1.24
pyyaml>=6.0
